remove obvious dependencies on association.to
trying to simplify the API for an association. the to property was always convenient for retrieving the table name easily, but forces a clunky API to the developer, where they have to keep the table names in lock-step with the models they pass as the second argument. This also gets sloppy in the polymorphic belongs to case, where you have to pass multiple tables in an array in lock-step with the modelCB array.

In a second commit, I will remove the api itself and see if any specs break from that. If so, I will repair them as well.