add balloon models, update has many and has one type signatures
Doing this because in order to load STI associations from a has one or has many relationship, we will need to know which models are available to marshal our payloads with.

Though the app is not entirely stable, all tests are still passing despite significant changes made to the has one and has many type signatures. I am anticipating this is because we have already handled much of the work needed to adapt surrounding types, given that belongs to associations already support arrays of models as a type.

In a successive commit, I will begin implementing the load pattern for has many and has one associations associations implementing sti.