rename query -> kyselyQuery where it makes sense
Doing this so that we can gain a somewhat reliable understanding of when our variable is pointing to a Query instance, and when it is pointing to a Kysely query instance.

Note: I left out many places where renaming the variable would have had to also change the argument name to the method.