Switch from `x.constructor === Array` to `Array.isArray(x)`
Was not able to reproduce in specs, but on a deployed
server, passing in `[]` resulted in `x.constructor === Array`
returning `false`. When the exception printed the constructor
name, it was "Array", and `JSON.stringify(x)` was `[]`, so
we don't understand how `x.constructor === Array` was
returning `false`. Switching to `Array.isArray(x)` is safely
reliable