fix client schema generator bug
when attempting to read a serializer with an association pointing to a
non-existent serializer (i.e. a model with no serializer getter
defined), an exception is thrown notifying us to define a default
serializer for that model. Rather than raise an exception in these
cases, we should be assigning an `any` type to those fields.

close https://rvohealth.atlassian.net/browse/PDTC-5062