add paramSafeColumns
In order to get the types right here, I was forced to add foreignKey to
our schema const, which then created problems because kysely blows away
and rewrites schema every time you run migrations and then we use schema
to determine what columns a model has, so there is a brief period where
our app is unable to produce foreign keys. This is easily fixed by
running the schema builder twice, but that is a hacky fix. I have left
detailed comments in the cli.ts file describing a potential fix to this.