catch exceptions on serializers getter
We wanted these errors to flow through originally, but we also
considered it a valid pattern to have a serializers getter that throws
an exception when accessed. The problem is, when syncing client api
schema, we can put ourselves in a situation where we often have to call
the serializers getter to determine what the corresponding serializer to
a model is.