HasMany/HasOne through is incompatible with many options since the through association is really meta-data for leveraging another association. All a through association can do is define the source and modify the query (e.g. where, whereNot)
Modify the types so that if `through` is an option,
then `source` and `preloadThroughColumns` are also
options, but many other things are not, and if one
of those things is an option, then `through` and
`source` are not.

Also define `distinct` where the other options are defined
and remove it from the HasOne options.