rewire associations to use string refs tied to global names
This enables us to dodge circular dependency issues as our apps grow in
complexity. We are beginning to encounter this now in one of our apps,
and we don't want to create and encourage design patterns that force
people into hard corners with associations because of javascript's
terrbile problems with circular dependencies.