raise hard exception if where clause receives undefined val
In javascript, it is considered valid to have an object whos keys are
pointing to undefined values. Prior to this PR, we would simply filter
out any keys which had undefined values. However, this could lead to
completely unexpected consequences in the case where you are not
anticipating that a value might be undefined, since providing undefined
as a value would simply omit that field. This led to a terrible problem
in a specific case for us, where we called `await User.findBy({ rvoId })`,
not anticipating that if rvoId were undefined, that it would just find
the first user in the db.

To protect against this easy mistake, this PR introduces new behavior
for underlying where clauses, to detect and raise an exception when
  encountering an undefined value, rather than simply filtering it out.

close https://rvohealth.atlassian.net/browse/PDTC-6152