remove support for ordered HasOne associations
This would create unexpected results when attempting to load an ordered
HasOne through a HasMany parent. The order would apply to the HasOne
association, and then a limit statement would sneak in to suppress all
but the first found record, which would cause unexpected behavior when
attempting to load multiple HasOne associations through the various
HasMany parent associations.