Reorganize decorators
All instance and static method decorators must be referenced from the `Deco` pattern. This allows for easy discovery of what decorators are available, reduces confusion over which decorators require the `Deco` pattern and which don't, and keeps the option open for future typing of arguments based on the Dream model.

Also remove Encrypt#hash/verifyHash because it isn't actually used by Dream, the currently recommended hashing algorithm (argon2) is not available in node crypto, and to properly support hashing, we'd want to introduce the legacy pattern present in #decrypt so that people could keep up with the currently recommended algorithm, but unlike with encryption, it isn't enough to simply support two (two with encryption enables one to iterate over all encrypted fields and re-encrypt) since hashing is one-way, so one would need to support the entire history of hashing algorithm changes used over time.