leverage AST within typescript to improve schema builder (#614)
The basic premise of this PR is to update all places in our sync process
where we are spitting out manually-formatted data into strings and
writing it to files. This is largely being done in 3 places:

1.) when enhancing the kysely-codegen output for usage within dream
2.) when generating the dream.ts schema files
3.) when generating the dream.globals.ts file

to approach this, a base class, `ASTBuilder`, was introduced to the code
base. It taps into low-level typescript factory utils and provides
helper functions to empower the child classes to compose with a little
more clarity when dealing with AST nodes in typescript, since this can
otherwise be incredibly hairy to read and comprehend. On top of this
class, three new child classes have been introduced, each to solve the
above 3 problem spots in our syncing:

1.) the `ASTKyselyCodegenEnhancer` class is used to update the generated
db.ts file after kysely-codegen sync, and handles all the enhancements
that were otherwise made by the original enhancement functions, which
have now been removed
2.) the SchemaBuilder class has been replaced with the
`ASTSchemaBuilder` class, which essentially outputs the same thing,
though it is now enhanced slightly:
  * no more DBClass, since DB covered the bases already
* prettier is now used (safely) to format output of all new schema files
3.) the `ASTGlobalSchemaBuilder` was introduced to replace the existing
global type building mechanisms

close https://rvohealth.atlassian.net/browse/PDTC-9446