add package manager check, along with ability to bypass with env
a.) this check used to exist within the psychic initialization, but then
dream absorbed the packageManager option, and did not inherit the check
along with it. This check is being moved from psychic to here.

b.) in create-psychic, we need to be able to write tests to try out new
package managers, but to do that safely, we need a way to bypass the
package manager check via env, so that create-psychic specs can
temporarily override it to audition new package managers.